+++ readme.md
@@ -154,6 +154,7 @@ Here is a human friendly list of them :
 |                                  |             |                                                                                                                                                                                          |
 | NegativeTravelTime               | Warning     | The travel duration between two stops is negative.                                                                                                                                       |
 | NegativeStopDuration             | Warning     | The `departure_time` at a stop is earlier than its `arrival_time`.                                                                                                                       |
+| MissingAgencyId                  | Error       | When there are multiple agencies, the `agency_id` must be filled for each route.                                                                                                           |
 | MissingName                      | Error       | An agency, a route or a stop has its name missing.                                                                                                                                       |
 | MissingCoordinates               | Warning     | A shape point or a stop is missing its coordinate(s).                                                                                                                                    |
 | NullDuration                     | Warning     | The travel duration between two stops is null.                                                                                                                                           |
