+++ src/issues.rs
@@ -98,6 +98,8 @@ pub enum IssueType {
     /// A trip must visit more than one stop in stop_times.txt to be usable by passengers for
     /// boarding and alighting.
     UnusableTrip,
+    /// When there are multiple agencies, `agency_id` must be filled for each route
+    MissingAgencyId,
 }
 
 /// Represents an object related to another object that is causing an issue.
