+++ src/validators/mod.rs
@@ -9,7 +9,7 @@ pub mod file_presence;
 pub mod interpolated_stoptimes;
 pub mod invalid_reference;
 pub mod raw_gtfs;
-pub mod route_type;
+pub mod routes;
 pub mod shapes;
 pub mod stop_times;
 pub mod stops;
