+++ test_data/missing_agency_id/agency.txt
@@ -1,0 +1,3 @@
+agency_id,agency_name,agency_url,agency_timezone,agency_lang
+a_1,"BIBUS",http://www.bibus.fr,Europe/Paris,fr
+a_2,"Ter",http://www.sncf.com,Europe/Paris,fr
