+++ test_data/missing_agency_id/calendar.txt
@@ -1,0 +1,7 @@
+service_id,monday,tuesday,wednesday,thursday,friday,saturday,sunday,start_date,end_date
+s_0,0,0,0,0,0,1,1,20170101,20180101
+s_1,0,0,0,0,0,1,1,20170101,20190101
+s_2,0,0,0,0,0,1,1,20160101,20170101
+s_3,0,0,0,0,0,1,1,20200101,20210101
+s_4,0,0,0,0,0,1,1,20210101,20220101
+s_5,0,0,0,0,0,1,1,20190101,20200101
