+++ test_data/missing_agency_id/calendar_dates.txt
@@ -1,0 +1,4 @@
+service_id,date,exception_type
+s_0,20230101,1
+s_0,20100101,2
+s_5,20300101,2
