+++ test_data/missing_agency_id/routes.txt
@@ -1,0 +1,4 @@
+route_id,agency_id,route_short_name,route_long_name,route_desc,route_type,route_url,route_color,route_text_color
+r_0,a_1,"100","100","",3,,000000,FFFFFF
+r_1,a_1,"100","100","",3,,000000,FFFFFF
+r_2,,"100","100","",3,,000000,FFFFFF
