+++ test_data/missing_agency_id/trips.txt
@@ -1,0 +1,7 @@
+route_id,service_id,trip_id,trip_headsign,trip_short_name,direction_id,block_id,wheelchair_accessible,bikes_allowed,trip_desc,shape_id
+r_1,s_0,close_stops,"85088452",,0,,0,0,,
+r_1,s_1,near_stops,"85088452",,0,,0,0,,
+r_1,s_2,null_duration,"85088452",,0,,0,0,,
+r_2,s_3,too_fast,"85088452",,0,,0,0,,
+r_2,s_4,too_slow,"85088452",,0,,0,0,,
+r_2,s_5,time_traveler,"85088452",,0,,0,0,,
